requirements.txt
Django==4.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
black==24.3.0
flake8==7.0.0
coverage==7.2.1
isort==5.13.2
